Moving a few things around
