--- client/app/reader/LoadingScreen/LoadingScreenActions.js
@@ -1,8 +1,0 @@
-import * as Constants from '../constants';
-
-export const setLoadedVacolsId = (vacolsId) => ({
-  type: Constants.SET_LOADED_APPEAL_ID,
-  payload: {
-    vacolsId
-  }
-});
