+++ client/app/reader/PdfViewer/actionTypes.js
@@ -18,3 +18,4 @@ export const SCROLL_TO_SIDEBAR_COMMENT = 'SCROLL_TO_SIDEBAR_COMMENT';
 
 export const RECEIVE_APPEAL_DETAILS = 'RECEIVE_APPEAL_DETAILS';
 export const RECEIVE_APPEAL_DETAILS_FAILURE = 'RECEIVE_APPEAL_DETAILS_FAILURE';
+export const SET_LOADED_APPEAL_ID = 'SET_LOADED_APPEAL_ID';
