+++ client/app/reader/constants.js
@@ -19,7 +19,6 @@ export const SET_SEARCH_INDEX = 'SET_SEARCH_INDEX';
 export const SET_SEARCH_INDEX_TO_HIGHLIGHT = 'SET_SEARCH_INDEX_TO_HIGHLIGHT';
 export const SET_DOCUMENT_LOAD_ERROR = 'SET_DOCUMENT_LOAD_ERROR';
 export const CLEAR_DOCUMENT_LOAD_ERROR = 'CLEAR_DOCUMENT_LOAD_ERROR';
-export const SET_LOADED_APPEAL_ID = 'SET_LOADED_APPEAL_ID';
 
 export const RECEIVE_APPEAL_DETAILS = 'RECEIVE_APPEAL_DETAILS';
 export const RECEIVE_APPEAL_DETAILS_FAILURE = 'RECEIVE_APPEAL_DETAILS_FAILURE';
